Rename RecipeCreatorUITests to RecipeCreatorViewUITests to avoid confusion Add RecipeCreatorUITests
